+++ src/main/java/com/furnitureapp/factory/clientele/CustomerFactory.java
@@ -5,7 +5,7 @@ import com.furnitureapp.utility.Helper;
 
 public class CustomerFactory {
     public static Customer createCustomer(String fullName, String emailAddress, String phoneNumber, String address, String password){
-        int customerCode = Helper.generateCode();
+        Long customerCode = Helper.generateCode();
 
         return new Customer.CustomerBuilder()
                 .setCustomerCode(customerCode)
